Home feed created
Layout for each meal item was made and the recycler view for the whole home feed was made as well. An adapter for the recycler view was made and set to the recycler view in MainActivity. The api was queried for random recipes which was used to populate the home feed. Each recipe's number of likes, servings and time to prepare was gotten and displayed on the home screen as well.