Meal and Post tables referenced in App
Created the Post and Meal classes in which the Post and Meal databases are referenced. Getters and setters are set in them to allow information from the database be used in the app and also allow the app to write to the database. The Post database will serve as the source of data for user reviews and comments to various meals and the  Meal database will hold the various meals which are obtained from the api and which users can comment on.