Similar Meals as Selected meals
The api has an endpoint that returns a list of similar recipes as the selected recipe. Here I have implemented retrofit interface and method call to return the required data. I have also implemented a listener that deals with the successful return of data or unsuccessful return of data. I have also added a recycler view which will display the similar meals in the recipe details view.