Recipe Reviews Implementation (#28)
* Test

* Recipe Reviews Implementation

A button has been created on the recipe details page which when clicked fires an intent to the Comments Activity which allows users to leave a review on a recipe

* Comment Page layout made

* Reviews saved to Parse Database

a user is able to make a review by typing texts and taking a picture of their finished meal and these data is stored in the Parse Database. Subsequently, these will be queried to populate a comments feed page which will include reviews left for a particular recipe.

* Comments on recipes displayed

Users are able to click a "see reviews" icon and be able to see a list of reviews that include photos of that particular meal.